#16 DataActor for User ~ Adjusted Behavior to new SQL stores specifications ~ fixes for UserDataBehavior ~ reworked permission SQL query

+++ app/src/main/scala/io/agamis/fusion/core/actors/data/entities/ProfileDataBehavior.scala
@@ -30,8 +30,8 @@ object ProfileDataBehavior {
     isActive: Option[Boolean],
     createdAt: List[(String, Instant)],
     updatedAt: List[(String, Instant)],
-    limit: Long,
-    offset: Long,
+    limit: Int,
+    offset: Int,
     orderBy: List[(String, Int)]
   ) extends Identifiable with Timetracked with Pageable
