+++ app/src/main/scala/io/agamis/fusion/core/actors/data/entities/common/Pageable.scala
@@ -1,7 +1,7 @@
 package io.agamis.fusion.core.actors.data.entities.common
 
 trait Pageable {
-  def limit: Long
-  def offset: Long
+  def limit: Int
+  def offset: Int
   def orderBy: List[(String, Int)]
 }
