+++ app/src/main/scala/io/agamis/fusion/external/api/rest/dto/user/UserQuery.scala
@@ -6,15 +6,15 @@ import java.time.Instant
 final case class UserQuery (
   id: List[String],
   username: List[String],
-  limit: Long,
-  offset: Long,
+  limit: Int,
+  offset: Int,
   createdAt: List[(String, Instant)],
   updatedAt: List[(String, Instant)],
   orderBy: List[(String, Int)]
 ) extends QueryBase(limit, offset, createdAt, updatedAt)
 
 object UserQuery {
-  def apply(param: (List[String], List[String], Long, Long, List[(String, String)], List[(String, String)], List[(String, Int)])): UserQuery = {
+  def apply(param: (List[String], List[String], Int, Int, List[(String, String)], List[(String, String)], List[(String, Int)])): UserQuery = {
     UserQuery(
       param._1,
       param._2,
